SS14-26964 Clown Waddling Replicates, etc
A bunch of cleanup and refactors to make clown waddling
generally better; fixed replication and some remaining bugs.
