+++ Content.Server/Movement/Systems/ServerWaddleAnimationSystem.cs
@@ -1,0 +1,5 @@
+﻿using Content.Shared.Movement.Systems;
+
+namespace Content.Server.Movement.Systems;
+
+public sealed class ServerWaddleAnimationSystem : SharedWaddleAnimationSystem;
